+++ sephora-backend/Application/Services/Interfaces/IProductService.cs
@@ -3,8 +3,8 @@
 public interface IProductService
 {
     Task<IEnumerable<ProductDto>> Get();
-    Task<ProductDto?> GetById(int id);
+    Task<ProductDto?> GetById(long id);
     Task Create(CreateProductDto createProductDto);
     Task Edit(EditProductDto editProductDto);
-    Task Delete(int id);
+    Task Delete(long id);
 }
